move permissions to shared, add new lint/fmt rules
