+++ .oxlintrc.jsonc
@@ -130,6 +130,7 @@
         "fixStyle": "separate-type-imports",
       },
     ],
+    "import/consistent-type-specifier-style": ["error", "prefer-top-level"],
   },
   "overrides": [
     {
