+++ backend/functions/src/routes/status.ts
@@ -1,19 +1,19 @@
-import type { Request, Response } from "express";
-import { Router } from "express";
-import { hasRoles, isAuthenticated } from "../middleware/authentication";
-import { db } from "..";
-import type { CollectionReference } from "firebase-admin/firestore";
-import { ReviewStatus } from "@app-portal/shared/constants";
+import { ReviewStatus, PermissionRole } from "@app-portal/shared/constants";
 import { DecisionConfirmationSchema } from "@app-portal/shared/types";
 import type {
   DecisionConfirmation,
   InternalApplicationStatus,
 } from "@app-portal/shared/types";
+import type { Request, Response } from "express";
+import { Router } from "express";
+import type { CollectionReference } from "firebase-admin/firestore";
+import { logger } from "firebase-functions";
+
+import { db } from "..";
+import { hasRoles, isAuthenticated } from "../middleware/authentication";
+import { validateSchema } from "../middleware/validation";
 import type { ApplicationForm } from "../models/appForm";
 import type { ApplicationResponse } from "../models/appResponse";
-import { logger } from "firebase-functions";
-import { validateSchema } from "../middleware/validation";
-import { PermissionRole } from "../models/user";
 
 const router = Router();
 const APPLICATION_STATUS_COLLECTION = "app-status";
