+++ frontend/src/components/dor/ReviewerAssignmentsDashboard/useRows.ts
@@ -1,15 +1,16 @@
+import { PermissionRole } from "@app-portal/shared/constants";
+import type { ApplicantRole } from "@app-portal/shared/constants";
 import type {
   AppReviewAssignment,
   ApplicationReviewData,
 } from "@app-portal/shared/types";
-import type { ApplicantRole } from "@app-portal/shared/constants";
+import { useQuery } from "@tanstack/react-query";
+
 import { getApplicationForm } from "@/services/applicationFormsService";
 import { getReviewDataForAssignment } from "@/services/reviewDataService";
 import { getUserById } from "@/services/userService";
 import type { ApplicantUserProfile } from "@/types/types";
-import { PermissionRole } from "@/types/types";
 import { calculateReviewScore } from "@/utils/scores";
-import { useQuery } from "@tanstack/react-query";
 
 export type AssignedAppRow = {
   index: number;
